Add license information to file listings
This small patch adds license_digest_list and license_list to each entry in a
directory listing. The field 'license_digest_list', confusingly, can have the
value 'OPEN' which does not translate to any license. Otherwise it contains
a list of hashes of the licenses required to download the corresponding file.
The field 'license_list' contains resolved License models (useful as they have
the 'theme' field which is user-readable)