I think this is what was causing recent problems on snapshots (unicode in URLs):
https://docs.djangoproject.com/en/1.3/ref/unicode/
URI and IRI handling
Web frameworks have to deal with URLs (which are a type of IRI). One requirement of URLs is that they are encoded using only ASCII characters. However, in an international environment, you might need to construct a URL from an IRI – very loosely speaking, a URI that can contain Unicode characters. Quoting and converting an IRI to URI can be a little tricky, so Django provides some assistance.

Have appled the supplied conversion function. Will test once on staging (the dev server doesn't suffer from the problem).