add support for DB migrations
This is a bit of a pain because we need to deal with our deployed
DB's that were never started with south. A bit of background might
be useful on how you handle this:

 http://stackful-dev.com/adding-south-migrations-to-your-django-project-post-factum.html

This new task will run during updates 1 time to provide the fake
migration we need. Given the fact our production services will not
have run syncdb for the new APILog table, I had to make that look
like a migration so that it will get created.

I also cleaned up a couple of pep8/pyflakes warning in settings.py

Change-Id: Ia65a00c8e2160d111335dcc6b291de801025afb8