Add /login redirect URL which will let user to login and is easy to remember.
Django admin dosn't seem to support external auth, like OpenID/Crowd that we
use. So, to access admin, one needs to pre-auth. With l-l-p, workaround to
do that is to find restricted build somewhere to trigger auth process.
Doesn't scale at all. Also, by wading thru sources, it turns out that
/linaro-openid/login/ URL can be used to login, but who can remember that?
So, just add easy-to-remember /login URL, symmetrical to /logout that we
have.

Change-Id: I8158aa90950ec34d03c1df9b8b35383765ca5149