bug #947: don't allow file overwrites in v2 api
This removes the ability to overwrite a file in the v2 apis. It
removes the ability to overwrite the build-info file in v1 while
logging api warnings for other overwrites

This change required me to address a long standing issue with the
v1 api test. The test needs a skeleton directory of files to drive
the testing, but also was *altering* that directory. This keeps
the skeleton directory read-only and copies things over to a temp
directory we can do whatever operation we choose on.

Change-Id: I2bdb9ce8ab977e9fd8b672411a48acd28bb24bdc