cleanup dir_list function (create Artifact object)
This function has grown too complex to manage. As we move to
S3 we need to deal with artifacts more generically (ie they could be
local or remote). This is the beginnings of that abstraction that we can
build on.

The Artifact object splits up all the logic that was done by dir_list
into smaller methods that can share instance data. Future patches will
convert logic for things like getting the build-info file for an artifact
or determining if the file is protected.

Change-Id: I245c3b31e1642e1ef43170f4f8b71e9416c301fc