move render-files logic into artifact
needed for s3

also fixed a string encoding issue with our how-to sample. the
file couldn't render unless your default code-page was ISO-8859. I've
made it UTF-8 to fit with everything else.

This can be checked in production at (look for tabbed help text):

 http://releases.linaro.org/15.06/android/images/reference-lcr-juno

and in the local sampleroot at:

 ~linaro-android/staging-snowball/173

Change-Id: Ia2f5ab0fc2e2bcd317273e49590d2996263e3ea3