move get_header_html_content to artifact object
needed for s3 support. This moves the logic into the BaseArtifact
object requiring the implementing class to only have to return the
contents of a given file in its directory.

This also required moving the test cases over to the common module.

This is hand testable with the sampleroot's "readme" directory and shows
no visible changes, and can be seen in production at:

 http://releases.linaro.org/15.06/ubuntu/boards/lt-d01

Change-Id: I8ed78d3bd081815afb82959f9a6741a01cc4e369