Add better handling of "path" for wildcard downloads
The current logic for knowing the "path" of a wildcard download works
okay, but leads to awkward reporting. For example a download of:

 /blah/latest/build*.img

will create a download entry for "build123.img". The actual download
*should* get reported as blah/latest/build123.img. This can be seen in
production on builds.96boards.org. The fix is fairly simple: rather than
using os.path.split, we just subtract the "basepath" from the full path
to determine its name.

Change-Id: I1f2d2f7d896563d4d085b451abcc0b84c1843f1d
Reviewed-on: https://review.linaro.org/9633
Reviewed-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>