reports: handle IPv6 addresses
We occasionally see an IPv6 address come in via an IPv4 tunnel. ie:

 HTTP_X_FORWARDED_FOR = ipv6
 REMOTE_ADDR = the ipv4 proxy

We don't have an IPV6 capability lookup database and they don't seem
that accurate. Furthermore, these don't currently happen often enough to
be statistically relevant. So we'll just fill in ipv6-unknown for now
and can come back at a later date if we want to try and fix.

Change-Id: Ib18430638dd6484680ac2f95e57f6cb5859163a6