reports: handle next_month logic properly
I found releases.linaro.org/reports couldn't render today because of an
error trying to find the "next_month" logic for a download on Jan 30.
The problem is it was trying to build a date of Feb 30 and Feb only has
28 (or 29) days. This fixes the issue by adding logic to understand the
days in a month and also adds some minimal testing.

Change-Id: I0bedb908d3fa63e1ae7eb3232284ccb451677a11
Reviewed-on: https://review.linaro.org/10242
Reviewed-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Reviewed-by: Ben Copeland <ben.copeland@linaro.org>