add logging for request ip misses
The logic for determining the actual request IP continues to be
unusually complicated. We've started getting "bad" IPs over the last few
days from builds.96boards.org. Somehow we are getting an ip of "rt504"
inserted into the DB which then causes errors when we try to get the GEO
IP information. This tries to validate we are storing a "real" IP, and
if we hit something we don't expect we log it, so we can try and get an
idea why its happening.

Change-Id: I0bb36062b434c18651138b0121a415bc0bd98ead
Reviewed-on: https://review.linaro.org/11004
Reviewed-by: Andy Doan <andy.doan+gerrit@linaro.org>