add support for HEAD requests
This is contingent on a version of mod_wsgi that supports HEAD (for
Ubuntu 16.04 and later). For local files, we have nothing to do as the
apache sendfile support is smart enough. For s3 we have to generate the
proper temp-url for the caller.

One thing that stinks with this change is that S3 can't be tested
properly with the Django client (it won't follow the redirect to S3).

Change-Id: I943c07e444aaf2db668cfa0e8726542449e3159a
Reviewed-on: https://review.linaro.org/14485
Reviewed-by: Andy Doan <andy.doan@linaro.org>