fix s3 urlencoded paths
We have some publishing jobs that are now naming files with ~'s in them.
The problem is that the tilde has to be converted to %7E so that its a
safe url-encoded string for S3 to sign. This leaves us calling the file
2 different things which S3 won't allow when trying to upload to the
signed url.

The safe way to handle this is to store all S3 files with urlencoded
names. We then modify our presentation layer to decode that name so its
listed sanely.

Change-Id: I782de89b25776d55ed03afa882245e0d01c5216b
Reviewed-on: https://review.linaro.org/19091
Reviewed-by: Ben Copeland <ben.copeland@linaro.org>
Reviewed-by: Fathi Boudra <fathi.boudra@linaro.org>