fix last commit
The last commit caused a big issue with *existing* files in S3 that had
URL encodable characters. Specifically the "+" character is fine in S3,
doesn't need to be encoded, and the change breaks doing lookups of it.

This change converts things for us to build up a known list of
substitiuions.

Change-Id: I2a1abbac6ec63aa9612196a7aba4f3be5ac4b1b1