download report: just skip bad download entries
We get these about once a week where a single download entry will have
some invalid utf8 junk in it. Its almost always just a single row, so
lets just discard it for now. We'll still get a single email so we could
in theory add the entry into the DB if we think its legit.

Change-Id: Icb8eebcae0f7b3f04756a3e22ea447a07a210bc2
Reviewed-on: https://review.linaro.org/20505
Reviewed-by: Ben Copeland <ben.copeland@linaro.org>