LLP: Changes for Django 1.11
Code changes to get LLP to 1.11. A bunch of features had to be
re-written due to Django deprecating features. A re-write of the url
modules had to be done, where we now need to import the views. We no
longer need to define templates all over the place, just one central
point and the test_views needed to be updated due to the Location header
not having the server URL.

Change-Id: I58f11a230423c9e521a54df3e0e65c72840a73e4
Reviewed-on: https://review.linaro.org/23481
Reviewed-by: Benjamin Copeland <ben.copeland@linaro.org>
Reviewed-by: Kelley Spoon <kelley.spoon@linaro.org>