Settings: Move report csv to prevent overwrites in docker
Because our docker design allows us to run multiple
containers that all share the same project directory,
we need a way to distinguish report CSVs and their lock
files running on different hosts.

This change moves the report CSV up one directory (which
will be /srv in docker and /srv/<container> on the host).

Change-Id: I451ea1c454be00e0f98f897ff450c6197c7b001f
Reviewed-on: https://review.linaro.org/25394
Reviewed-by: Benjamin Copeland <ben.copeland@linaro.org>