s3_purge: fix the releases deletemarker bug and code cleanup
When testing this the releases/ bucket wasn't checking that the
deletemarker was being set. Obviously this is very dangerous so a fix
has been put in to ensure a deletemarker *is* set before deleting the
file.

Since we handle the same call a couple of times, I have put
handle_permanent_deletes, into its own function which has reduced a lot
of reputation.

Along with those fix, the code has been cleaned up to be a bit more
readable

Change-Id: Iffd2f55d04bc889050bfd7a19e6f62fe205de0c6
Reviewed-on: https://review.linaro.org/29085
Reviewed-by: Benjamin Copeland <ben.copeland@linaro.org>