s3_purge: enable delete_key to set delete marker
Previously the s3_purge script had the line to
delete an object commented out.  Since the
publishing buckets are using versioning, this
seems to have prevented setting the DeleteMarker
on a file object.

As a result, a check in handle_permanent_deletes() never
removed anything since nothing had DeleteMarker set, and
old versions of files were allowed to stack up.

Change-Id: I34524d3b0b7776d9d41603d8e0e9f34ef1b526ec
Reviewed-on: https://review.linaro.org/c/infrastructure/linaro-license-protection/+/32999
Reviewed-by: Kelley Spoon <kelley.spoon@linaro.org>