s3_flatten: add s3_flatten script
This script is intended to replace the s3_purge command.

It simplifies the logic by breaking the clean up process
into 3 parts:
 1 - "flatten" all file objects to have no more than 1
     version object and optionally 1 deletemarker regardless
     of prefix

 2 - delete any expired files from prefixes that have not been
     excluded via S3_PURGE_EXCLUDES

 3 - place deletemarkers on files that are older than MARK_DAYS
     that have not been excluded via S3_PURGE_EXCLUDES

The other difference with this script is that it attempts to
determine what objects to delete as it iterates through the
versioned bucket listing, and will collect multiple delete
operations into a single multidelete request.

Change-Id: I25a227f574504bb4e828f29670aec366e87952c9
Reviewed-on: https://review.linaro.org/c/infrastructure/linaro-license-protection/+/33268
Reviewed-by: Benjamin Copeland <ben.copeland@linaro.org>