llp: make delete for link_latest aware of bucket versioning
Currently LLP will call Bucket.list() on a bucket to
generate a list of keys to be deleted from the target
directory of the link_latest() API function.

This can cause a problem if the underlying bucket is
using S3 versioning as the S3.Key information returned
by Bucket.list() does not include version information.
S3 will interpret this as a request to insert deletemarkers
at the top of all keys in the list rather than actually
removing them, and this can dramatically increase the object
count of the bucket.

This change will check to see if a bucket is using versioning,
and if so use the Bucket.list_versions() call in order to make
sure all S3 objects in the target directory are removed.

Change-Id: Ifb14aaa7b4166c83741abcd98339d36e72b5ee5c
Reviewed-on: https://review.linaro.org/c/infrastructure/linaro-license-protection/+/33579
Reviewed-by: Riku Voipio <riku.voipio@linaro.org>
Reviewed-by: Benjamin Copeland <ben.copeland@linaro.org>