s3_flatten.py:  add retries for boto3 bucket operations
It turns out that this script can occasionally error out
if the AWS API doesn't respond with proper HTTP.  When
this happens, the error isn't caught and the script ends
prematurely and can leave large amounts of unnecessary
data in the bucket.

This wraps S3 bucket operations (specifically delete_keys
and list_objectS) in a try statement that will attempt
up to 3 retries after sleeping for 30 seconds.

Change-Id: I3f4c945937b20635166cad093630bfbca9026a78
Reviewed-on: https://review.linaro.org/c/infrastructure/linaro-license-protection/+/35063
Reviewed-by: Benjamin Copeland <ben.copeland@linaro.org>