s3_flatten.py: wrap the entire handle_bucket() method in retry logic
It turns out that boto will attempt to lazily load details for objects
in a bucket.  As a result, it is not sufficient to simply wrap the
list_objects() calls.  In order to prevent even more complexity
by attempting to predict where try/excepts retries are needed,
let's just wrap the handle_bucket() method itself and use the
wrapper to safely clean up objects.

Change-Id: I257f9390dbb6f59e83511e50594824306ad9baf8
Reviewed-on: https://review.linaro.org/c/infrastructure/linaro-license-protection/+/35184
Reviewed-by: Kelley Spoon <kelley.spoon@linaro.org>