Add timestamps to the report CSV
Currently, our process_report script assumes that a Download occured
on the day the process_report script was run.  This isn't correct,
so we need to keep track of the actual datetime when the download
occurred in the CSV as an extra field.

This attempts to handle it in a way that is backwards compatible with
previous CSV files that don't include a timestamp column.

Change-Id: I5249c434a2777f09fb1e2010b812a91255517b07
Reviewed-on: https://review.linaro.org/c/infrastructure/linaro-license-protection/+/36874
Reviewed-by: Kelley Spoon <kelley.spoon@linaro.org>