settings.py: add in testplan and testreport changes
A while back a request was made to include a 'Test Plan'
and 'Test Report' tab for some of our projects that would
feature the contents of the respective files.  Apparently
the change was made in place on the server, but was never
migrated into the LLP repo.

Let's get caught up.

Change-Id: I5e49fcd41211966065edbc01191ad607d84e00c5
Reviewed-on: https://review.linaro.org/c/infrastructure/linaro-license-protection/+/38759
Reviewed-by: Kelley Spoon <kelley.spoon@linaro.org>