manage.py: silence beautifulsoup warnings
BeautifulSoup produces and UserWarning when imported reminding the user
that the version we're using is EOL and that we should be upgrading it
soon.

This warning is producing output that is causing crontab to email when
there are no errors or problems to report.  Let's silence it until
we can finally update the software.

Change-Id: Ic45819c0f22ff91b2f2b5138433d69428cbeec95
Reviewed-on: https://review.linaro.org/c/infrastructure/linaro-license-protection/+/40042
Reviewed-by: Riku Voipio <riku.voipio@linaro.org>