staticfiles: get control of the static files once again
Fix staticfiles so that they are correctly generated by
'collectstatic' regardless of where or how this app
is deployed.

In this change, let's concentrate all static files in
'license_protected_downloads/static' so that we can
clear out STATIC_ROOT ('static') and rebuild it reliably
with 'collectstatic'.

We also restore the contents of the
'license_protected_downloads/static' directory, and update
the unit-test.sh script to call 'collectstatic' on a freshly
installed STATIC_ROOT before it runs.

Signed-off-by: Kelley Spoon <kelley.spoon@linaro.org>
Change-Id: Ice720b4794a48540c3753f4a6efba027cbc29040
Reviewed-on: https://review.linaro.org/c/infrastructure/linaro-license-protection/+/40206