staticfiles: rename root static dir and add missing files
By adding 'static' to the gitignore file, the pattern match was
also applied to 'license-protected-downloads/static' and the
missing admin, jquery, and css files were left out of the
repo.

Let's rename the top level directory where the collected files are
to be put to 'staticroot' so that it can safely be added to
the gitignore file and restore the missing files.

Change-Id: Ie051df5c26ab83df96af5a8453648b0db9f3cbcd
Reviewed-on: https://review.linaro.org/c/infrastructure/linaro-license-protection/+/40226
Reviewed-by: Kelley Spoon <kelley.spoon@linaro.org>