llp: remove crowd and openid support in favor of ldap auth
In order to remove the final piece of Crowd usage in our
infrastructure, let's migrate LLP to using ldap for both
authentication and group look ups.

We also need to update the views, urls, and tests to
no longer rely upon crowd/openid and instead mock up
the LDAP lookups.

Change-Id: I254d7bbffff97ffa8ebe5f5be72e763e77a6c45d
Signed-off-by: Kelley Spoon <kelley.spoon@linaro.org>
Reviewed-on: https://review.linaro.org/c/infrastructure/linaro-license-protection/+/40275