group_auth_ldap: fix an error with user group authorization
There is a bug in the ldap group authorization code
where we use the full django username for authentication
(which is the full email), but only the UID (first.lastname)
is stored in the group membership table.

We should also take this time to just try to look up
the UID in the groups table instead of trying to build
a list of group memberships for the user and compare
that to the required groups.

Change-Id: I41209fb8745a6225f3e7344910dc89c19d336a76
Signed-off-by: Kelley Spoon <kelley.spoon@linaro.org>
Reviewed-on: https://review.linaro.org/c/infrastructure/linaro-license-protection/+/40448